requirements.txt
bilibili-api-python
loguru
openai
APScheduler
openai-whisper
pytest
httpx